+++ .eslintrc
@@ -41,7 +41,8 @@
 		"prefer-spread": 2,
 		"operator-linebreak": ["error", "before"],
 		"array-bracket-spacing": ["error", "never", { "singleValue": false }],
-		"object-curly-spacing": ["error", "always"]
+		"object-curly-spacing": ["error", "always"],
+		"no-import-assign": 0
 	},
 	"overrides": [
 		{
