+++ src/js/utils.js
@@ -41,4 +41,32 @@ function roundBits(number) {
 	return (number * 1000000 + 0.5 | 0) / 1000000
 }
 
-export { timeString, roundBits };
+function compareArr(arr, out) {
+	let minX = 1000
+	let maxX = -1000
+	let minY = 1000
+	let maxY = -1000
+	let minZ = 1000
+	let maxZ = -1000
+	let num = 0
+	for (let i = 0; i < arr.length; i += 3) {
+		num = arr[i]
+		minX = minX > num ? num : minX
+		maxX = maxX < num ? num : maxX
+		num = arr[i + 1]
+		minY = minY > num ? num : minY
+		maxY = maxY < num ? num : maxY
+		num = arr[i + 2]
+		minZ = minZ > num ? num : minZ
+		maxZ = maxZ < num ? num : maxZ
+	}
+	out[0] = minX
+	out[1] = minY
+	out[2] = minZ
+	out[3] = maxX
+	out[4] = maxY
+	out[5] = maxZ
+	return out
+}
+
+export { timeString, roundBits, compareArr };
